feat(girth): private key was decrypted
Refs: #1667
