+++ packages/core/src/external/saml/xca/iti38-envelope.ts
@@ -230,9 +230,10 @@ export function createITI38SoapEnvelope({
 export function createAndSignDQRequest(
   bodyData: DQBodyData,
   publicCert: string,
-  privateKey: string
+  privateKey: string,
+  privateKeyPassword: string
 ): string {
   const xmlString = createITI38SoapEnvelope({ bodyData, publicCert });
-  const fullySignedSaml = signFullSaml({ xmlString, publicCert, privateKey });
+  const fullySignedSaml = signFullSaml({ xmlString, publicCert, privateKey, privateKeyPassword });
   return fullySignedSaml;
 }
